# Run OrderedSink before PredicatedSink to improve performance