This change was benchmarked with contactDAO. contactDAO was populated with 50k entries, and then a 1000 iterations of the following query were run for each trial:
```
 contactDAO
      .orderBy(new foam.mlang.order.Desc(Contact.ID))
      .limit(500)
      .select(new ArraySink());
```

Results were as follows:

### Prior to the change
trial1: 7.8ms avg
trial2: 7.1ms avg
trial3: 8.8ms avg

### After the change
trial1: 6.7ms avg
trial2: 5.8ms avg
trial3: 6.7ms avg