@KernelDeimos 2019-11-05
This needs more evaluation - it resulted in a performance increase when ordering by ID, but not when ordering by other fields