@kgrgreer 2019-11-05
In your benchmark, is contactDAO predicated? Or is it authenticated, which is providing the predicate?