@kgrgreer 2019-11-05
I think we can make this better by only reversing the order if there is a Limit. Otherwise, the old order will be more efficient.