@KernelDeimos 2019-11-11
I'm actually not so sure there's a performance increase anymore. There should be, but I wasn't able to observe it again after I fixed a problem with my benchmarking.

I'm going to record what I learned in this comment so it doesn't get lost:

## Benchmarking
### Preconditions:
- contactDAO with 1000 records (close to the number in production)
- user is a non-admin Ablii user (so that authorization predicate is required)
- query specifies order for a property other than ID (FIRST_NAME was used)
- query has limit and skip (90 and 750 were used arbitrarily)

### Results after moving order (this PR)
- About 10-15% slower; couldn't figure out why

### Results after adding sink to create fallback sub when sub is null
- No significant different observer. I want to double-check this result later.

## Manual testing
- Manually loading a contactDAO with a very high number of records (50,000 was used) results in a very slow pageload because getting the record count takes a long time. Getting the 90 records for the table takes at most a couple seconds.