@kgrgreer 2019-11-11
If the predicate is expensive and you have a limit, then it makes sense to order first because then you won't need to apply the predicate to entries that appear after the limit.
However, if you don't have a limit, then you'll need to apply the predicate to all entries anyway, so it should be faster to predicate first so that you have fewer entries to sort.
So, the order should depend on if you have limit or not.
Also, it could depend on whether you have a "slow" or "fast" predicate, but that's probably not worth accounting for (in terms of code complexity).