@TW80000 2019-11-19
I spent some time yesterday testing this out. I ran into an issue, unfortunately. When I switched the order so that the predicate gets applied after ordering, I got an error saying `User` couldn't be cast to `Contact` (which is a subclass of User and is stored in the same DAO). The cast was happening in a call to `.compare`, which was being called as part of the sort that `OrderedSink` does. This makes sense, since `contactDAO` is just a `FilteredDAO` on top of `userDAO`, which just adds a predicate. But since that predicate gets applied _after_ the order when we switch the order of the sink decorators, the order is still being applied to other objects in the DAO that aren't `Contact`s.

Due to this issue, I don't think it's possible to do the predicate after ordering without making additional changes.

The thought crossed my mind of supporting a way to flag predicates to denote whether they should apply before or after the sort, but it's a bit of an ugly idea. We'd have to walk the predicate if it contains `AND`s or `OR`s and look for instances of predicates that apply before or after, then split it up and apply some before and the rest after, while still keeping the desired behaviour and semantics of the original predicate.

Another thought is that it might be safe to say that `IsInstanceOf` and `IsClassOf` are special cases and should always be applied before the order. You'd still have to walk the predicate to extract these "special" predicates though. I'm not certain that these two predicates are always special cases though, or whether others are special cases as well.